refactor(db): extract schema migrations
